refactor: API 변경사항 반영
- 기존 API 주소를 유지
- 변경사항 반영
